--- src/main/scala/com/fortysevendeg/github4s/app.scala
@@ -1,9 +1,0 @@
-package com.fortysevendeg.github4s
-
-import cats.data.Coproduct
-import com.fortysevendeg.github4s.free.algebra._
-
-object app {
-  type COGH01[A] = Coproduct[RepositoryOp, UserOp, A]
-  type GitHub4s[A] = Coproduct[AuthOp, COGH01, A]
-}
