--- src/main/scala/com/fortysevendeg/github4s/free/domain/Pagination.scala
@@ -1,6 +1,0 @@
-package com.fortysevendeg.github4s.free.domain
-
-case class Pagination(
-  page: Int,
-  per_page: Int
-)
