--- src/main/scala/com/fortysevendeg/github4s/free/domain/Repository.scala
@@ -1,6 +1,0 @@
-package com.fortysevendeg.github4s.free.domain
-
-case class Repository(
-  id: Int,
-  name: String
-)
