--- src/main/scala/com/fortysevendeg/github4s/free/domain/User.scala
@@ -1,10 +1,0 @@
-package com.fortysevendeg.github4s.free.domain
-
-case class User(
-  id: Int,
-  login: String,
-  avatar_url: String,
-  html_url: String,
-  name: Option[String] = None,
-  email: Option[String] = None
-)
