+++ src/main/scala/github4s/free/domain/Pagination.scala
@@ -1,0 +1,6 @@
+package github4s.free.domain
+
+case class Pagination(
+  page: Int,
+  per_page: Int
+)
