+++ src/main/scala/github4s/free/domain/Repository.scala
@@ -1,0 +1,6 @@
+package github4s.free.domain
+
+case class Repository(
+  id: Int,
+  name: String
+)
