+++ src/main/scala/github4s/free/domain/User.scala
@@ -1,0 +1,10 @@
+package github4s.free.domain
+
+case class User(
+  id: Int,
+  login: String,
+  avatar_url: String,
+  html_url: String,
+  name: Option[String] = None,
+  email: Option[String] = None
+)
